Extract lazy lists tests.
